v4l: Add sub-stream mask id for buffer type
To support multiple sub-streams on a single video
node (and DMA engine), allow multiplexing the (buffer)
type field on the structs it is present. Add a flag to
struct video_device to tell that a driver support
sub-streams which is off by default, and take the flag
into account in check_fmt().

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>